backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Supabase client
supabase==2.0.0

# Environment and configuration
python-dotenv==1.0.0

# Data validation and serialization
pydantic[email]==2.5.0

# Password hashing
bcrypt==4.1.2
PyJWT==2.8.0

# HTTP client (for Supabase)
httpx==0.25.2

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1

# Type hints support
typing-extensions==4.8.0

# Date and time utilities (built-in, but explicit for clarity)
# datetime - built-in
# hashlib - built-in
# secrets - built-in
# hmac - built-in
# base64 - built-in
# os - built-in
# sys - built-in
# time - built-in
# threading - built-in
# uuid - built-in
# logging - built-in
# contextlib - built-in
# argparse - built-in

# Development dependencies (optional)
# black==23.11.0
# flake8==6.1.0
# mypy==1.7.1